added effect calcs, dummy data and visuals
Decided to do the allocation, selection, and interaction effects in-sheet for now. Added random dummy data and a few visuals for readability. Felt like I was overcomplicating the issue by trying to do all of it in Python. There will still be use for the Python programme; other visualisation, importing data from elsewhere to overwrite the sheet, etc